refactor(preprocess): makes CLI more POSIX
- outputs on stdout
- more options
- doc accessible from the exe
